Move trade interface to strata-basics
Add new interface FinanceTrade in strata-finance
Rename tradeInfo() method to info()
