+++ examples/src/main/java/com/opengamma/strata/examples/finance/CdsTradeExample.java
@@ -8,10 +8,10 @@ package com.opengamma.strata.examples.finance;
 
 import com.google.common.collect.ImmutableList;
 import com.google.common.io.CharStreams;
+import com.opengamma.strata.basics.Trade;
 import com.opengamma.strata.collect.io.ResourceLocator;
 import com.opengamma.strata.examples.marketdata.credit.markit.MarkitRedCode;
 import com.opengamma.strata.examples.report.TradePortfolio;
-import com.opengamma.strata.finance.Trade;
 import com.opengamma.strata.finance.credit.CdsTrade;
 import com.opengamma.strata.finance.credit.type.CdsConventions;
 import org.joda.beans.ser.JodaBeanSer;
