+++ modules/finance/src/main/java/com/opengamma/strata/finance/impl/loader/BulletPaymentFpmlTradeParser.java
@@ -5,8 +5,8 @@
  */
 package com.opengamma.strata.finance.impl.loader;
 
+import com.opengamma.strata.basics.Trade;
 import com.opengamma.strata.collect.io.XmlElement;
-import com.opengamma.strata.finance.Trade;
 import com.opengamma.strata.finance.TradeInfo;
 import com.opengamma.strata.finance.loader.FpmlDocument;
 import com.opengamma.strata.finance.loader.FpmlTradeParser;
