+++ modules/finance/src/main/java/com/opengamma/strata/finance/loader/FpmlTradeParser.java
@@ -8,11 +8,11 @@ package com.opengamma.strata.finance.loader;
 import org.joda.convert.FromString;
 import org.joda.convert.ToString;
 
+import com.opengamma.strata.basics.Trade;
 import com.opengamma.strata.collect.ArgChecker;
 import com.opengamma.strata.collect.io.XmlElement;
 import com.opengamma.strata.collect.named.ExtendedEnum;
 import com.opengamma.strata.collect.named.Named;
-import com.opengamma.strata.finance.Trade;
 
 /**
  * Pluggable FpML trade parser.
