+++ modules/pricer/src/main/java/com/opengamma/strata/pricer/calibration/CalibrationDerivative.java
@@ -7,9 +7,9 @@ package com.opengamma.strata.pricer.calibration;
 
 import java.util.List;
 
+import com.opengamma.strata.basics.Trade;
 import com.opengamma.strata.collect.array.DoubleArray;
 import com.opengamma.strata.collect.array.DoubleMatrix;
-import com.opengamma.strata.finance.Trade;
 import com.opengamma.strata.market.curve.definition.CurveParameterSize;
 import com.opengamma.strata.math.impl.function.Function1D;
 import com.opengamma.strata.pricer.rate.ImmutableRatesProvider;
