+++ modules/pricer/src/main/java/com/opengamma/strata/pricer/calibration/CalibrationMeasure.java
@@ -5,7 +5,7 @@
  */
 package com.opengamma.strata.pricer.calibration;
 
-import com.opengamma.strata.finance.Trade;
+import com.opengamma.strata.basics.Trade;
 import com.opengamma.strata.market.sensitivity.CurveCurrencyParameterSensitivities;
 import com.opengamma.strata.pricer.rate.ImmutableRatesProvider;
