+++ modules/pricer/src/main/java/com/opengamma/strata/pricer/calibration/CalibrationMeasures.java
@@ -13,8 +13,8 @@ import java.util.List;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.ImmutableSet;
+import com.opengamma.strata.basics.Trade;
 import com.opengamma.strata.collect.array.DoubleArray;
-import com.opengamma.strata.finance.Trade;
 import com.opengamma.strata.market.curve.definition.CurveParameterSize;
 import com.opengamma.strata.market.sensitivity.CurveCurrencyParameterSensitivities;
 import com.opengamma.strata.market.sensitivity.CurveCurrencyParameterSensitivity;
