+++ modules/report/src/main/java/com/opengamma/strata/report/framework/expression/ResultsRow.java
@@ -10,6 +10,7 @@ import static com.opengamma.strata.collect.Guavate.toImmutableList;
 import java.util.List;
 import java.util.Set;
 
+import com.opengamma.strata.basics.Trade;
 import com.opengamma.strata.collect.result.FailureReason;
 import com.opengamma.strata.collect.result.Result;
 import com.opengamma.strata.collect.type.TypedString;
@@ -18,7 +19,6 @@ import com.opengamma.strata.engine.config.Measure;
 import com.opengamma.strata.finance.Product;
 import com.opengamma.strata.finance.ProductTrade;
 import com.opengamma.strata.finance.SecurityTrade;
-import com.opengamma.strata.finance.Trade;
 import com.opengamma.strata.function.StandardComponents;
 import com.opengamma.strata.report.ReportCalculationResults;
