+++ modules/report/src/main/java/com/opengamma/strata/report/framework/expression/TradeTokenEvaluator.java
@@ -14,7 +14,8 @@ import org.joda.beans.JodaBeanUtils;
 import org.joda.beans.MetaBean;
 
 import com.google.common.collect.Sets;
-import com.opengamma.strata.finance.Trade;
+import com.opengamma.strata.basics.Trade;
+import com.opengamma.strata.finance.FinanceTrade;
 import com.opengamma.strata.finance.TradeInfo;
 
 /**
@@ -32,7 +33,11 @@ public class TradeTokenEvaluator extends TokenEvaluator<Trade> {
   @Override
   public Set<String> tokens(Trade trade) {
     MetaBean metaBean = JodaBeanUtils.metaBean(trade.getClass());
-    return Sets.union(metaBean.metaPropertyMap().keySet(), trade.getTradeInfo().propertyNames());
+    if (trade instanceof FinanceTrade) {
+      FinanceTrade ftrade = (FinanceTrade) trade;
+      return Sets.union(metaBean.metaPropertyMap().keySet(), ftrade.getTradeInfo().propertyNames());
+    }
+    return metaBean.metaPropertyMap().keySet();
   }
 
   @Override
@@ -53,16 +58,19 @@ public class TradeTokenEvaluator extends TokenEvaluator<Trade> {
     }
 
     // trade info
-    Optional<String> tradeInfoPropertyName = trade.getTradeInfo().propertyNames().stream()
-        .filter(p -> p.equalsIgnoreCase(firstToken))
-        .findFirst();
+    if (trade instanceof FinanceTrade) {
+      FinanceTrade ftrade = (FinanceTrade) trade;
+      Optional<String> tradeInfoPropertyName = ftrade.getTradeInfo().propertyNames().stream()
+          .filter(p -> p.equalsIgnoreCase(firstToken))
+          .findFirst();
 
-    if (tradeInfoPropertyName.isPresent()) {
-      Object propertyValue = trade.getTradeInfo().property(tradeInfoPropertyName.get()).get();
+      if (tradeInfoPropertyName.isPresent()) {
+        Object propertyValue = ftrade.getTradeInfo().property(tradeInfoPropertyName.get()).get();
 
-      return propertyValue != null ?
-          EvaluationResult.success(propertyValue, remainingTokens) :
-          EvaluationResult.failure("Property '{}' not set", firstToken);
+        return propertyValue != null ?
+            EvaluationResult.success(propertyValue, remainingTokens) :
+            EvaluationResult.failure("Property '{}' not set", firstToken);
+      }
     }
     return invalidTokenFailure(trade, firstToken);
   }
