+++ modules/report/src/test/java/com/opengamma/strata/report/framework/expression/TradeTokenEvaluatorTest.java
@@ -13,9 +13,9 @@ import org.testng.annotations.Test;
 
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableSet;
+import com.opengamma.strata.basics.Trade;
 import com.opengamma.strata.collect.id.StandardId;
 import com.opengamma.strata.finance.SecurityLink;
-import com.opengamma.strata.finance.Trade;
 import com.opengamma.strata.finance.TradeInfo;
 import com.opengamma.strata.finance.future.GenericFuture;
 import com.opengamma.strata.finance.future.GenericFutureTrade;
